+++ IteaIOSDip/Models/Courses/CourseDecription/DetailedCourseViewController.swift
@@ -19,7 +19,6 @@ class DetailedCourseViewController: UIViewController {
     @IBOutlet weak var enrollButton: UIButton!
     @IBOutlet weak var courseProgramButton: UIButton!
     
-    
     var course = Courses()
     var recieveCourse = Courses()
     
@@ -28,10 +27,11 @@ class DetailedCourseViewController: UIViewController {
         
         course = recieveCourse
         detailInfoCourse(course: course)
-        
-        updateDetailUI()
-        visualUpdateView(view: skillsView)
-        visualUpdateView(view: courseAboutView)
+        backgroundView(image: "backgroundfor", color: UIColor.init(red: 19/255, green: 58/255, blue: 85/255, alpha: 1), alpha: 0.9)
+        courseAboutView.blurDetailView()
+        skillsView.blurDetailView()
+        enrollButton.iteaButton()
+        courseProgramButton.iteaButton()
     }
     
     @IBAction func didTapGoBackButton(_ sender: Any) {
@@ -41,113 +41,24 @@ class DetailedCourseViewController: UIViewController {
     @IBAction func didTapCourseProgramActionButton(_ sender: Any) {
         let storyboard = UIStoryboard(name: "Courses", bundle: nil)
         let viewController = storyboard.instantiateViewController(withIdentifier: "ProgramCourseViewController") as! ProgramCourseViewController
-        
         viewController.courseRecieve = course
-        
         navigationController?.pushViewController(viewController, animated: true)
-        
-        
     }
+    
     @IBAction func didTapCourseSignInActionButton(_ sender: Any) {
-        
         let storyboard = UIStoryboard(name: "CourseSignUp", bundle: nil)
-               let viewController = storyboard.instantiateViewController(withIdentifier: "CourseSignUpViewController") as! CourseSignUpViewController
-        
+        let viewController = storyboard.instantiateViewController(withIdentifier: "CourseSignUpViewController") as! CourseSignUpViewController
         viewController.recieveCourse = course
-        
-               navigationController?.pushViewController(viewController, animated: true)
+        navigationController?.pushViewController(viewController, animated: true)
     }
 }
 
+//  MARK: - UPDATE INFO -
 extension DetailedCourseViewController {
-    
-    
     func detailInfoCourse(course: Courses) {
-        
         courseNameTextLabel.text = course.courseName
         aboutCourseTextLabel.text = course.courseDescription
         skillsTextLabel.text = course.courseSkills
         flowHeaderImageView.image = UIImage(named: course.courseHeaderImage ?? "noHeaderImage")
-        
     }
-    
-    func visualUpdateView(view: UIView) {
-          
-          view.clipsToBounds = true
-          view.layer.masksToBounds = false
-          view.layer.cornerRadius = 12
-          view.layer.shadowColor = UIColor.white.cgColor
-          view.layer.shadowOffset = CGSize(width: 3, height: 3)
-          view.layer.shadowRadius = 10.0
-          view.layer.shadowOpacity = 0.5
-
-      }
-    
-    func updateDetailUI() {
-          
-//          let backgroundImage = UIImageView(frame: UIScreen.main.bounds)
-//          backgroundImage.image = UIImage(named: "bt")
-//          backgroundImage.contentMode = .scaleAspectFill
-//          let backgroundForImage = UIView(frame: UIScreen.main.bounds)
-//          backgroundForImage.backgroundColor = UIColor.black
-//          backgroundForImage.alpha = 0.9
-//          view.addSubview(backgroundImage)
-//          view.addSubview(backgroundForImage)
-//        self.view.sendSubviewToBack(backgroundImage)
-//          self.view.sendSubviewToBack(backgroundForImage)
-          
-        let backgroundImage = UIImageView(frame: UIScreen.main.bounds)
-               let backgroundForImage = UIView(frame: UIScreen.main.bounds)
-               backgroundImage.image = UIImage(named: "backgroundfor")
-               backgroundImage.contentMode = .scaleAspectFill
-               backgroundForImage.backgroundColor = UIColor.init(red: 19/255, green: 58/255, blue: 85/255, alpha: 1)
-               backgroundForImage.alpha = 0.9
-               view.addSubview(backgroundImage)
-               view.addSubview(backgroundForImage)
-               self.view.sendSubviewToBack(backgroundForImage)
-               self.view.sendSubviewToBack(backgroundImage)
-          
-        let blurEffect = UIBlurEffect(style: .prominent)
-                 let blurView = UIVisualEffectView(effect: blurEffect)
-                 blurView.translatesAutoresizingMaskIntoConstraints = false
-                 courseAboutView.insertSubview(blurView, at: 0)
-             
-                 
-                 NSLayoutConstraint.activate([blurView.heightAnchor.constraint(equalTo: courseAboutView.heightAnchor), blurView.widthAnchor.constraint(equalTo: courseAboutView.widthAnchor)
-                 ])
-                 let secondBlurView = UIVisualEffectView(effect: blurEffect)
-                 secondBlurView.translatesAutoresizingMaskIntoConstraints = false
-                 skillsView.insertSubview(secondBlurView, at: 0)
-                
-                 NSLayoutConstraint.activate([secondBlurView.heightAnchor.constraint(equalTo: skillsView.heightAnchor), secondBlurView.widthAnchor.constraint(equalTo: skillsView.widthAnchor),
-                 ])
-        
-        
-        enrollButton.clipsToBounds = true
-         
-            enrollButton.layer.cornerRadius = 12.0
-            enrollButton.isOpaque = true
-            enrollButton.alpha = 0.8
-            enrollButton.layer.masksToBounds = false
-            enrollButton.layer.shadowColor = UIColor.black.cgColor
-            enrollButton.layer.shadowOpacity = 0.5
-            enrollButton.layer.shadowOffset = CGSize(width: 10, height: 10)
-            enrollButton.layer.shadowRadius = 12.0
-        
-      
-        
-           courseProgramButton.layer.cornerRadius = 12.0
-           courseProgramButton.isOpaque = true
-           courseProgramButton.alpha = 0.8
-           courseProgramButton.layer.masksToBounds = false
-           courseProgramButton.layer.shadowColor = UIColor.black.cgColor
-           courseProgramButton.layer.shadowOpacity = 0.5
-           courseProgramButton.layer.shadowOffset = CGSize(width: 10, height: 10)
-           courseProgramButton.layer.shadowRadius = 12.0
-          
-      }
-    
-    
-    
 }
-
