+++ IteaIOSDip/Models/Courses/CoursesFlowCell/CoursesCollectionViewCell.swift
@@ -10,22 +10,17 @@ import UIKit
 
 class CoursesCollectionViewCell: UICollectionViewCell {
 
-    
     @IBOutlet weak var mainBackgroundView: UIView!
     @IBOutlet weak var viewCourseLogoView: UIView!
     @IBOutlet weak var courseLogoImageView: UIImageView!
-    
     @IBOutlet weak var mainCourseNameTextLabel: UILabel!
-    
     
     override func awakeFromNib() {
         super.awakeFromNib()
-        // Initialization code
     }
-
 }
 
-
+//  MARK: - UPDATE INFO CELL -
 extension CoursesCollectionViewCell {
     
     func updateCourseCellInfo(coursesFlow: CourseFlow) {
@@ -37,23 +32,14 @@ extension CoursesCollectionViewCell {
         courseLogoImageView.layer.cornerRadius = courseLogoImageView.frame.width / 2
         courseLogoImageView.layer.borderColor = UIColor.white.cgColor
         courseLogoImageView.layer.borderWidth = 3.0
-        
-//        mainBackgroundView.layer.backgroundColor =
-        
-        
         courseLogoImageView.clipsToBounds = true
         courseLogoImageView.layer.masksToBounds = false
-//        courseLogoImageView.layer.cornerRadius = 12
         courseLogoImageView.layer.shadowColor = UIColor.black.cgColor
         courseLogoImageView.layer.shadowOffset = CGSize(width: 3, height: 3)
         courseLogoImageView.layer.shadowRadius = 6
         courseLogoImageView.layer.shadowOpacity = 0.4
         courseLogoImageView.layer.backgroundColor = UIColor.random().cgColor
-        
-        
     }
-    
-    
 }
