+++ IteaIOSDip/Models/Courses/CoursesFlowViewController.swift
@@ -22,8 +22,7 @@ class CoursesFlowViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        uiDesignCoursesFlow()
+        backgroundView(image: "bg", color: UIColor.clear, alpha: 1)
         iteaStudent = recieveUserInfo
         
         if filterFlow.count == 0 {
@@ -31,7 +30,6 @@ class CoursesFlowViewController: UIViewController {
         } else {
             flow = filterFlow
         }
-    
         coursesCollectionView.register(UINib(nibName: "CoursesCollectionViewCell", bundle: nil), forCellWithReuseIdentifier: "CoursesCollectionViewCell")
         coursesCollectionView.reloadData()
     }
@@ -49,33 +47,4 @@ class CoursesFlowViewController: UIViewController {
         navigationController?.pushViewController(viewController, animated: true)
     }
 }
-
-//  MARK: - COLLECTION VIEW - ALL COURSES -
-extension CoursesFlowViewController: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout{
-    
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 1
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return flow.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "CoursesCollectionViewCell", for: indexPath) as! CoursesCollectionViewCell
-        cell.updateCourseCellInfo(coursesFlow: flow[indexPath.row])
-        return cell
-    }
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let size = CGSize(width: collectionView.frame.width / 2 - 5, height: 200 )
-           return size
-    }
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let storyboard = UIStoryboard(name: "Courses", bundle: nil)
-        let viewController = storyboard.instantiateViewController(withIdentifier: "FlowViewController") as! FlowViewController
-        viewController.recieveFlow = flow[indexPath.row]
-        navigationController?.pushViewController(viewController, animated: true)
-    }
-}
-
