+++ IteaIOSDip/Models/Courses/CoursesViewController+Extension.swift
@@ -8,19 +8,38 @@
 
 import UIKit
 
-//  MARK: - UI DESIGN / BACKGROUND / ETC
-extension CoursesViewController {
+//  MARK: - TABLE VIEW DELEGATE -
+extension CoursesViewController: UITableViewDelegate, UITableViewDataSource {
     
-    func uiDesignCourses() {
-        let backgroundImage = UIImageView(frame: UIScreen.main.bounds)
-        let backgroundForImage = UIView(frame: UIScreen.main.bounds)
-        backgroundImage.image = UIImage(named: "backgroundfor")
-        backgroundImage.contentMode = .scaleAspectFill
-        backgroundForImage.backgroundColor = UIColor.init(red: 19/255, green: 58/255, blue: 85/255, alpha: 1)
-        backgroundForImage.alpha = 0.9
-        view.addSubview(backgroundImage)
-        view.addSubview(backgroundForImage)
-        self.view.sendSubviewToBack(backgroundForImage)
-        self.view.sendSubviewToBack(backgroundImage)
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return 1
+    }
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return courses.count
+    }
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "CoursesTableViewCell") as! CoursesTableViewCell
+        cell.updateCourseInfo(course: courses[indexPath.row])
+        cell.delegate = self
+        cell.tag = indexPath.row
+        return cell
+    }
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        return 60
+    }
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: false)
+    }
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        let view = UIView(frame: CGRect(x: 0, y: 0, width: tableView.frame.width, height: 40))
+        let label = UILabel(frame: view.frame)
+        label.text = flow.name ?? ""
+        label.textAlignment = .center
+        label.textColor = UIColor.white
+        label.layer.backgroundColor = UIColor(red: 19/255, green: 58/255, blue: 85/255, alpha: 1).cgColor
+        label.font = UIFont.boldSystemFont(ofSize: 20.0)
+        label.text = label.text?.uppercased()
+        view.addSubview(label)
+        return view
     }
 }
