+++ IteaIOSDip/Models/Courses/ProgramingFlowCell/ProgrammingFlowTableViewCell.swift
@@ -11,23 +11,15 @@ import UIKit
 class ProgrammingFlowTableViewCell: UITableViewCell {
 
     @IBOutlet weak var programmingFlowImageView: UIImageView!
-    
     @IBOutlet weak var programmingFlowTextlabel: UILabel!
-    
-    
-    
+
     override func awakeFromNib() {
         super.awakeFromNib()
-        // Initialization code
     }
-
- 
     
     func updateFlowCourseInfo(programmingFlow: TypesOfProgrammingFlow) {
-        
         programmingFlowTextlabel.text = programmingFlow.name
         programmingFlowImageView.image = UIImage(named: programmingFlow.flowImage ?? "noImage")
-        
         programmingFlowImageView.contentMode = .scaleAspectFit
         programmingFlowImageView.clipsToBounds = true
         programmingFlowImageView.layer.masksToBounds = false
@@ -36,8 +28,5 @@ class ProgrammingFlowTableViewCell: UITableViewCell {
         programmingFlowImageView.layer.shadowOffset = CGSize(width: 3, height: 3)
         programmingFlowImageView.layer.shadowRadius = 12.0
         programmingFlowImageView.layer.shadowOpacity = 0.3
-
-        
     }
-    
 }
